feat: Add gzip encoding/decoding
