+++ internal/handlers/handlers_test.go
@@ -2,7 +2,6 @@ package handlers
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/go-resty/resty/v2"
 	"github.com/sergeysynergy/gopracticum/internal/storage"
 	"github.com/sergeysynergy/gopracticum/pkg/metrics"
 	"github.com/stretchr/testify/assert"
@@ -180,257 +179,6 @@ func TestList(t *testing.T) {
 
 			// повтороне закрытие тела ответа нужно чтобы автотест пропустил
 			resp.Body.Close()
-		})
-	}
-}
-
-func TestUpdate(t *testing.T) {
-	type want struct {
-		statusCode int
-	}
-	tests := []struct {
-		name string
-		body metrics.Metrics
-		want want
-	}{
-		{
-			name: "Not implemented",
-			body: metrics.Metrics{MType: "unknown"},
-			want: want{
-				statusCode: http.StatusNotImplemented,
-			},
-		},
-		{
-			name: "Gauge ok",
-			body: metrics.Metrics{
-				ID:    "Alloc",
-				MType: "gauge",
-				Value: func() *float64 { v := 42.24; return &v }(),
-			},
-			want: want{
-				statusCode: http.StatusOK,
-			},
-		},
-		{
-			name: "Counter ok",
-			body: metrics.Metrics{
-				ID:    "PollCount",
-				MType: "counter",
-				Delta: func() *int64 { d := int64(2); return &d }(),
-			},
-			want: want{
-				statusCode: http.StatusOK,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			handler := New()
-			r := chi.NewRouter()
-			r.Post("/update/", handler.Update)
-			ts := httptest.NewServer(r)
-			defer ts.Close()
-
-			client := resty.New()
-			resp, err := client.R().
-				EnableTrace().
-				SetHeader("Content-Type", "application/json").
-				SetBody(tt.body).
-				Post(ts.URL + "/update/")
-
-			assert.NoError(t, err)
-			assert.Equal(t, tt.want.statusCode, resp.StatusCode())
-		})
-	}
-}
-
-func TestUpdateHardBody(t *testing.T) {
-	type want struct {
-		statusCode int
-	}
-	tests := []struct {
-		name        string
-		contentType string
-		body        []byte
-		want        want
-	}{
-		{
-			name: "Update unsupported media type",
-			body: []byte(""),
-			want: want{
-				statusCode: http.StatusUnsupportedMediaType,
-			},
-		},
-		{
-			name:        "Update not acceptable",
-			body:        []byte("Not acceptable"),
-			contentType: "application/json",
-			want: want{
-				statusCode: http.StatusNotAcceptable,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			handler := New()
-			r := chi.NewRouter()
-			r.Post("/update/", handler.Update)
-			ts := httptest.NewServer(r)
-			defer ts.Close()
-
-			client := resty.New()
-			resp, err := client.R().
-				EnableTrace().
-				SetHeader("Content-type", tt.contentType).
-				SetBody(tt.body).
-				Post(ts.URL + "/update/")
-
-			assert.NoError(t, err)
-			assert.Equal(t, tt.want.statusCode, resp.StatusCode())
-			assert.Equal(t, "application/json", resp.Header().Get("Content-Type"))
-		})
-	}
-}
-
-func TestValueContentType(t *testing.T) {
-	h := New()
-	ts := httptest.NewServer(h.router)
-	defer ts.Close()
-
-	client := resty.New()
-	resp, err := client.R().
-		SetHeader("Accept", "application/json").
-		SetHeader("Content-Type", "bad type").
-		Post(ts.URL + "/value/")
-
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusUnsupportedMediaType, resp.StatusCode())
-	assert.Equal(t, "application/json", resp.Header().Get("Content-Type"))
-}
-
-func TestValueUnmarshalError(t *testing.T) {
-	h := New()
-	ts := httptest.NewServer(h.router)
-	defer ts.Close()
-
-	client := resty.New()
-	resp, err := client.R().
-		SetHeader("Accept", "application/json").
-		SetHeader("Content-Type", "application/json").
-		SetBody("{bad bad json").
-		Post(ts.URL + "/value/")
-
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusNotAcceptable, resp.StatusCode())
-	assert.Equal(t, "application/json", resp.Header().Get("Content-Type"))
-}
-
-func TestValue(t *testing.T) {
-	type myMetrics struct {
-		ID    string  `json:"id"`
-		MType string  `json:"type"`
-		Delta int64   `json:"delta,omitempty"`
-		Value float64 `json:"value,omitempty"`
-	}
-	type want struct {
-		statusCode int
-		body       myMetrics
-	}
-	tests := []struct {
-		name    string
-		handler *Handler
-		body    metrics.Metrics
-		want    want
-	}{
-		{
-			name:    "Metric type needed",
-			handler: New(),
-			body:    metrics.Metrics{MType: ""},
-			want: want{
-				statusCode: http.StatusBadRequest,
-			},
-		},
-		{
-			name:    "Not implemented",
-			handler: New(),
-			body:    metrics.Metrics{MType: "unknown"},
-			want: want{
-				statusCode: http.StatusNotImplemented,
-			},
-		},
-		{
-			name:    "Gauge not found",
-			handler: New(),
-			body: metrics.Metrics{
-				ID:    "Alloc",
-				MType: "gauge",
-			},
-			want: want{
-				statusCode: http.StatusNotFound,
-			},
-		},
-		{
-			name:    "Counter not found",
-			handler: New(),
-			body: metrics.Metrics{
-				ID:    "PollCount",
-				MType: "counter",
-			},
-			want: want{
-				statusCode: http.StatusNotFound,
-			},
-		},
-		{
-			name:    "Gauge ok",
-			handler: New(WithStorage(storage.New(storage.WithGauges(map[string]metrics.Gauge{"Alloc": 1221.23})))),
-			body: metrics.Metrics{
-				ID:    "Alloc",
-				MType: "gauge",
-			},
-			want: want{
-				statusCode: http.StatusOK,
-				body: myMetrics{
-					ID:    "Alloc",
-					MType: "gauge",
-					Value: 1221.23,
-				},
-			},
-		},
-		{
-			name:    "Counter ok",
-			handler: New(WithStorage(storage.New(storage.WithCounters(map[string]metrics.Counter{"PollCount": 42})))),
-			body: metrics.Metrics{
-				ID:    "PollCount",
-				MType: "counter",
-			},
-			want: want{
-				statusCode: http.StatusOK,
-				body: myMetrics{
-					ID:    "PollCount",
-					MType: "counter",
-					Delta: 42,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ts := httptest.NewServer(tt.handler.router)
-			defer ts.Close()
-
-			m := myMetrics{}
-			client := resty.New()
-			resp, err := client.R().
-				SetHeader("Accept", "application/json").
-				SetHeader("Content-Type", "application/json").
-				SetBody(tt.body).
-				SetResult(&m).
-				Post(ts.URL + "/value/")
-
-			assert.NoError(t, err)
-			assert.Equal(t, tt.want.statusCode, resp.StatusCode())
-			assert.Equal(t, "application/json", resp.Header().Get("Content-Type"))
-			assert.EqualValues(t, tt.want.body, m)
 		})
 	}
 }
