+++ internal/handlers/value.go
@@ -2,9 +2,10 @@ package handlers
 
 import (
 	"encoding/json"
-	"github.com/sergeysynergy/gopracticum/pkg/metrics"
 	"io/ioutil"
 	"net/http"
+
+	"github.com/sergeysynergy/gopracticum/pkg/metrics"
 )
 
 func (h *Handler) Value(w http.ResponseWriter, r *http.Request) {
